✨ 레시피 추천 페이지에 Swiper 구성 요소 및 스타일 상수 추가, UI 컴포넌트 분리
